requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
joblib
xgboost
twilio
python-dotenv
groq
pydantic